feat: redesign layout around Frame, Flex grow weights, and fixed-size Layer
